requirements.txt
google-generativeai
requests
# openai  # Opcional si decides no usar Gemini
